updating taskCtrl&taskService

+++ src/services/taskService.ts
@@ -2,8 +2,13 @@ import { TaskDB } from '../db/models/taskSchema';
 import { UserDB } from '../db/models/userSchema';
 import { AddTask, UpdateTask } from '../validation/task';
 
-export const addTask = async (data: AddTask) => {
-  const { title, description, dueDate, priority, isCompleted, tags, subtasks, userId } = data;
+export const addTask = async (userId: string, data: AddTask) => {
+  const user = await UserDB.findById(userId);
+  if (!user) {
+    return;
+  }
+
+  const { title, description, dueDate, priority, isCompleted, tags, subtasks } = data;
 
   return await TaskDB.create({
     title,
@@ -17,11 +22,19 @@ export const addTask = async (data: AddTask) => {
   });
 };
 
-export const updateTaskById = async (id: string, data: UpdateTask) => {
-  const { title, description, dueDate, priority, isCompleted, tags, subtasks } = data;
+export const updateTaskById = async (userId: string, id: string, data: UpdateTask) => {
+  const user = await UserDB.findById(userId);
+  if (!user) {
+    return;
+  }
 
-  return await TaskDB.findByIdAndUpdate(
-    id,
+  const { title, description, dueDate, priority, isCompleted, tags, subtasks } = data;
+  if (!TaskDB.find({ userId, _id: id })) {
+    return;
+  }
+
+  return await TaskDB.findOneAndUpdate(
+    { _id: id, userId },
     {
       title,
       description,
@@ -30,9 +43,24 @@ export const updateTaskById = async (id: string, data: UpdateTask) => {
       isCompleted,
       tags,
       subtasks,
+      userId,
     },
     { new: true, runValidators: true }
   );
+  // return await TaskDB.findByIdAndUpdate(
+  //   id,
+  //   {
+  //     title,
+  //     description,
+  //     dueDate,
+  //     priority,
+  //     isCompleted,
+  //     tags,
+  //     subtasks,
+  //     userId,
+  //   },
+  //   { new: true, runValidators: true }
+  // );
 };
 
 export const getTaskById = async (userId: string, id: string) => {
@@ -40,13 +68,25 @@ export const getTaskById = async (userId: string, id: string) => {
   if (!user) {
     throw new Error('User not found');
   }
-  return await TaskDB.find({ _id: id, userId });
+  const task = await TaskDB.findOne({ _id: id, userId });
+  if (!task) {
+    return;
+  }
+  return task;
 };
 
 export const getAllTasks = async (userId: string) => {
+  const user = await UserDB.findById(userId);
+  if (!user) {
+    throw new Error('User not found');
+  }
   return await TaskDB.find({ userId });
 };
 
-export const deleteTaskById = async (id: string) => {
-  return await TaskDB.findByIdAndDelete(id);
+export const deleteTaskById = async (userId: string, id: string) => {
+  const user = await UserDB.findById(userId);
+  if (!user) {
+    throw new Error('User not found');
+  }
+  return await TaskDB.findByIdAndDelete({ _id: id, userId });
 };
